+++ examples/blinky.rs
@@ -34,10 +34,10 @@ fn main() -> ! {
     let mut pwr = dp.PWR.constrain(&mut rcc.apb1r1);
 
     // Try a different clock configuration
-    let clocks = rcc.cfgr.hclk(8.mhz()).freeze(&mut flash.acr, &mut pwr);
+    let clocks = rcc.cfgr.hclk(8.MHz()).freeze(&mut flash.acr, &mut pwr);
     // let clocks = rcc.cfgr
-    //     .sysclk(64.mhz())
-    //     .pclk1(32.mhz())
+    //     .sysclk(64.MHz())
+    //     .pclk1(32.MHz())
     //     .freeze(&mut flash.acr);
 
     // let mut gpioc = dp.GPIOC.split(&mut rcc.ahb2);